Better Image API (rescale etc.) + fixed some Arrow issues (#97)

+++ include/sfw/Gfx/backend-specific/SFML/Box.hpp
@@ -60,7 +60,10 @@ public:
         sf::Vector2f itemSize = item.getSize();
         // Center item
         item.setPosition(
-            {getPosition().x + (size.x - itemSize.x) / 2, getPosition().y + (size.y - itemSize.y) / 2}
+            {(float)int(getPosition().x + (size.x - itemSize.x) / 2),
+             (float)int(getPosition().y + (size.y - itemSize.y) / 2)}
+            //! The weird back-and-forth casting is floor() without <cmath>
+            //! (see issue #97), and also to shut up the int->float warnings
         );
     }
